Personal Ai Knowledge assistant/requirements.txt
# Personal AI Knowledge Assistant - Dependencies
# Following LegalBot pattern with enhanced document processing

# Core Framework
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn>=0.24.0

# LLM Integration
groq>=0.4.0
openai>=1.3.0

# Embeddings and Vector Search
sentence-transformers>=2.2.2
scikit-learn>=1.3.0
numpy>=1.24.0

# Document Processing
PyPDF2>=3.0.1
python-docx>=0.8.11
python-magic>=0.4.27

# File Monitoring and Processing
watchdog>=3.0.0
python-magic-bin>=0.4.14

# Data Handling
pandas>=2.0.0
pydantic>=2.4.0

# Environment and Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Security and Encryption
cryptography>=41.0.0

# Logging and Utilities
colorama>=0.4.6
tqdm>=4.66.0

# Optional: Enhanced Document Processing
# Uncomment if you need these advanced features
# pytesseract>=0.3.10  # OCR for images in PDFs
# pillow>=10.0.0       # Image processing
# beautifulsoup4>=4.12.0  # HTML parsing
# lxml>=4.9.0          # XML/HTML parsing

# Development and Testing (Optional)
# pytest>=7.4.0
# black>=23.0.0
# flake8>=6.0.0
